[Labels] Validate labels (#7483)
Adding validation to verify that labels are a dictionary of strings.

Additionally, the maximum length for the label key is 255 characters.
[Here](https://github.com/mlrun/mlrun/pull/7149), we validate the label
value, but not the label key.
This causes the error: `Failed committing changes to DB`.

These bugs are relevant for all resources that have labels stored in a
different table:

- Artifact
- Run
- Schedule
- Function
- Project
- Feature
- Entity
- FeatureSet
- FeatureVector
- ModelEndpoint

https://iguazio.atlassian.net/browse/ML-9632
https://iguazio.atlassian.net/browse/ML-9599