[Application] Set multiple ports to runtime spec (#8392)
Jira - https://iguazio.atlassian.net/browse/ML-9752

Adds ability to set multiple ports to sidecar spec. Basically the idea
is that internal port will remain with its purpose - being a port to
which we redirect traffic through nuclio function. When
`application_port` parameter will define all the ports that application
should expose. Setters for these values contain special logic to make
sure that application port always has internal port in it.


In addition to that, added `port` to create_api_gateway in case
`direct_access` is True, so that user be able to create api gateway on
any of the ports of the deployed app.

Tested manually ✅