[API] Moved system id sync to online phase (#8500)
### 📝 Description

before the sync was on `_setup_service` which may be finished while
chief/worker is yet to be online as they are pending migration and not
all tables are ready yet.
now the sync is part of moving to online phase, where the data/schemas
are mgirated and db is ready to be used.

---

### 🛠️ Changes Made
<!-- - Key changes (e.g., added feature X, refactored Y, fixed Z) -->

---

### ✅ Checklist
- [ ] I updated the documentation (if applicable)
- [x] I have tested the changes in this PR

---

### 🧪 Testing
<!-- - How it was tested (unit tests, manual, integration) -->  
<!-- - Any special cases covered. -->  

patched vm, verified migration, from scratch and simple scale up/down to
see system id is returned on client-spec requests.

---

### 🔗 References
- Ticket link: https://iguazio.atlassian.net/browse/ML-10929
- Design docs links:
- External links:

---

### 🚨 Breaking Changes?

- [ ] Yes (explain below)
- [x] No

<!-- If yes, describe what needs to be changed downstream: -->

---

### 🔍️ Additional Notes
<!-- Anything else reviewers should know (follow-up tasks, known issues,
affected areas etc.). -->
<!-- ### 📸 Screenshots / Logs -->