[Application] Improve ports enrichment logic to preserve BC (#8516)
### 📝 Description

This pr fixes the bug created by this
[pr](https://github.com/mlrun/mlrun/pull/8392). The reason of the bug is
the new parameter spec.application_ports in application ports, that is
enriched on the __init__ and on setting the internal_application_port.
This parameter is responsible for setting all active ports of the
sidecar. So, on init:

```
internal_application_port = 8050
application_ports = [8050]
```
after setting internal application port to 5000:
```
internal_application_port = 5000
application_ports = [5000, 8050]
```
this can be workaround-ed by setting `spec.application_ports = [5000]`.
In order to preserve BC, If the application_ports list contains only one
port (the old internal application port), the list is rewritten to
include only the new internal application port. This ensures consistency
when the internal application port is updated. This ensures proper BC
with previous versions.

---

### ✅ Checklist
- [ ] I updated the documentation (if applicable)
- [x] I have tested the changes in this PR

---

### 🧪 Testing
tested on vmdev using the notebook from the ticket

---

### 🔗 References
- Ticket link:https://iguazio.atlassian.net/browse/ML-10806

---

### 🚨 Breaking Changes?

- [ ] Yes (explain below)
- [x] No