[Log Collector] Create file before finding pod (#8542)
### 📝 Description

When the log collector can't find the pod to stream logs from, it fails
and doesn't create a file.
If the run is deleted before the api calls `start_logs` on it again, it
falls back to the legacy log collection, [which does a k8s api call to
get the
logs](https://github.com/mlrun/mlrun/blob/3398b8ade56e6474cf195be28686fdc9804b7115/server/py/services/api/runtime_handlers/base.py#L1669-L1696).
In case we have many pods in this state, we are doing many k8s api call
which can choke the API.

To alleviate this a bit, we now first create an empty file for the run
in the log collector, before trying to find the pod.
The legacy log collection will not run, as it first checks if the file
exists.

---

### 🛠️ Changes Made
- Ensure the run has a log file before starting log collection.

---

### ✅ Checklist
- [ ] I updated the documentation (if applicable)
- [x] I have tested the changes in this PR

---

### 🧪 Testing
- All tests pass

---

### 🔗 References
- Ticket link: https://iguazio.atlassian.net/browse/ML-10765
- Design docs links:
- External links:

---

### 🚨 Breaking Changes?

- [ ] Yes (explain below)
- [x] No

---

### 🔍️ Additional Notes
<!-- Anything else reviewers should know (follow-up tasks, known issues,
affected areas etc.). -->
<!-- ### 📸 Screenshots / Logs -->