[Retry] Fix runs with hyperparams (#8535)
### 📝 Description
This PR fixes an issue with runs that include hyperparameters.
Previously, if a run failed and one or more of its tasks were set to
`pendingRetry`, those tasks were incorrectly treated as
`running`(=healthy). As a result, the entire run was marked as
`completed`.

With this fix, if any task is in the `pendingRetry` state, the run
itself will correctly be set to `pendingRetry`. And then be retried.

---


### ✅ Checklist
- [ ] I updated the documentation (if applicable)
- [x] I have tested the changes in this PR

---

### 🧪 Testing
Tested on vmdev:

<img width="2359" height="227" alt="image"
src="https://github.com/user-attachments/assets/a9518840-0653-43ec-90b8-87870cbdda56"
/>
First is the run which was executed on the image that contained the fix.
We can see that 3/3 retries happened, they failed and function was set
to error at the end of those. Between retries state was `pending`.

Second is the one on which I was reproducing the error. It is marked as
`completed` and we can see that it is stuck on 1/3 retries, even though
the run failed.

---

### 🔗 References
- Ticket link: https://iguazio.atlassian.net/browse/ML-10816

---

### 🚨 Breaking Changes?

- [ ] Yes (explain below)
- [x] No