[Runs] Allow running remote jobs while kfp is not set (#8600)
### 📝 Description

(taken from jira)
When the config mlrun.mlconf.force_run_local == "auto" (which is the
default), MLRun’s run_function determines if the job should run locally
or not by checking if the the value of mlrun.mlconf.kfp_urlis empty or
not.
Since KFP is disabled - the value is empty, and it is automatically run
locally.

---

### 🛠️ Changes Made

Moved the check to happen *only* when running a workflow.

---

### ✅ Checklist
- [ ] I updated the documentation (if applicable)
- [x] I have tested the changes in this PR

---

### 🧪 Testing

tested by adding unit-test coverage and debugging

---

### 🔗 References
- Ticket link: https://iguazio.atlassian.net/browse/ML-11036
- Design docs links:
- External links:

---

### 🚨 Breaking Changes?

- [ ] Yes (explain below)
- [x] No

<!-- If yes, describe what needs to be changed downstream: -->

---

### 🔍️ Additional Notes
<!-- Anything else reviewers should know (follow-up tasks, known issues,
affected areas etc.). -->
<!-- ### 📸 Screenshots / Logs -->