[Dependencies] Refactor Makefile and requirements for Python version handling (#8653)
### 📝 Description

Previous changes made on
https://github.com/mlrun/mlrun/pull/8463/files#diff-4d7c51b1efe9043e44439a949dfd92e5827321b34082903477fd04876edb7552
made newer protobuf (>3) into python 3.9 installation of mlrun which
then while using kfp 1.8.22 have failed due to conflicting dependencies.

---

### 🛠️ Changes Made

- Updated the `Makefile` to simplify the upgrade process for test
dependencies by removing specific Python version targets and
consolidating output files.
- Adjusted `requirements.txt` to specify version constraints for
`v3io-frames` based on Python version.
- Modified `dockerfiles/constraints-py3.9.txt` and
`dockerfiles/constraints-py3.11.txt` to align with the new versioning
strategy for `v3io-frames`.
- Updated various `locked-requirements.txt` files across different
dockerfiles to reflect the changes in dependency versions and ensure
compatibility with Python 3.9 and 3.11.


---

### ✅ Checklist
- [ ] I updated the documentation (if applicable)
- [x] I have tested the changes in this PR

---

### 🧪 Testing

- took a snapshot of 1.7.2 `pip freeze` installed deps on an old setup,
and based on that list, tried to upgrade to 1.10 to see it fails and
after my changes it suceeded
- built mlrun kfp images to see the installed kfp version is aligned
(1.8.22 to 3.9 and 1.8.23 to .311)


---

### 🔗 References
- Ticket link: https://iguazio.atlassian.net/browse/ML-11090
- Design docs links:
- External links:

---

### 🚨 Breaking Changes?

- [ ] Yes (explain below)
- [x] No

<!-- If yes, describe what needs to be changed downstream: -->

---

### 🔍️ Additional Notes
<!-- Anything else reviewers should know (follow-up tasks, known issues,
affected areas etc.). -->
<!-- ### 📸 Screenshots / Logs -->