[Serving] Add `ChoiceByField` Step (#8670)
### 📝 Description
The default `Choice` step by default sends the event to all the outlets
(children) of the step. Any other logic requires subclassing and adding
a custom `select_outlets` flow.

A more user-friendly approach would be for the `Choice` step to receive
a name of field from which to extract the route path. This way the user
would create a custom logic step that adds a field into the event with
the route name, and then just use the default choice step logic with
that field name.

---

### 🛠️ Changes Made
Adding ChoiceByField class that subclass `storey.Choice`.

---

### ✅ Checklist
- [X] I updated the documentation (if applicable)
- [X] I have tested the changes in this PR

---

### 🧪 Testing
- Unitest
---

### 🔗 References
- Ticket link: https://iguazio.atlassian.net/browse/ML-10097
- Design docs links:
- External links:

---

### 🚨 Breaking Changes?

- [] Yes (explain below)
- [X] No

---

### 🔍️ Additional Notes

---------

Co-authored-by: Gal Topper <gal.topper@gmail.com>