[Test] Normalize artifacts datetime fields prior to deepdiff comparison (#8732)
### 📝 Description
<!-- A short summary of what this PR does. -->
<!-- Include any relevant context or background information. -->
Test was failing transiently due to a difference in datetime object
formatting. In `datetime.isoformat()` if the microsecond value is 0,
Python omits it by default (documented
[here](https://docs.python.org/3/library/datetime.html#datetime.datetime.isoformat).
SQLAlchemy stores timestamps with microsecond precision, even if they
are `.0000`. This can sometimes cause a formatting mismatch, resulting
in the assertion error described in the ticket.
This PR normalize datetime fields prior to the deepdiff comparison



---

### 🛠️ Changes Made
<!-- - Key changes (e.g., added feature X, refactored Y, fixed Z) -->
- Added a function `_normalize_datetime_fields` inside
`_validate_run_artifacts` to handle the formatting mismatch

---

### ✅ Checklist
- [ ] I updated the documentation (if applicable)
- [x] I have tested the changes in this PR
- [ ] If I introduced a deprecation:
  - [ ] I followed the [Deprecation Guidelines](./DEPRECATION.md)
  - [ ] I updated the relevant Jira ticket for documentation

---

### 🧪 Testing

To test this I have altered the `_update_artifact_record_from_dict`
function which is responsible to update the `created` and `updated`
fields when storing an artifact. I made this function save the artifact
with hardcoded values for the datetime fields, making sure that the
changes in the PR solve them.

---

### 🔗 References
- Ticket link: https://iguazio.atlassian.net/browse/ML-11231
- Design docs links:
- External links:
https://docs.python.org/3/library/datetime.html#datetime.datetime.isoformat

---

### 🚨 Breaking Changes?

- [ ] Yes (explain below)
- [ ] No

<!-- If yes, describe what needs to be changed downstream: -->

---

### 🔍️ Additional Notes
<!-- Anything else reviewers should know (follow-up tasks, known issues,
affected areas etc.). -->
<!-- ### 📸 Screenshots / Logs -->