[Test] Fix system test "test_kfp_terminate_pipeline" (#8748)
### 📝 Description
<!-- A short summary of what this PR does. -->
<!-- Include any relevant context or background information. -->

This PR fixes the `test_kfp_terminate_pipeline` system test. This test
can only work with KFP 2+ (IGZ 3.7+) so the test is currently skipped
(explanation on why in this jira
https://iguazio.atlassian.net/browse/ML-10283) . Once system test env is
bumped to IGZ 3.7, it should be re-enabled.

The test was refactored to only check that the pipeline terminates. The
duplicate `terminate_pipeline` requests flow cannot be tested in a
system test, and it is already checked in unit test as part of #8013 in
`test_terminate_pipeline_success` test.

---

### 🛠️ Changes Made
<!-- - Key changes (e.g., added feature X, refactored Y, fixed Z) -->
- Refactor the `test_kfp_terminate_pipeline` system test so it works
with KFP 2+ (IGZ 3.7+). Test is currently skipped


---

### ✅ Checklist
- [ ] I updated the documentation (if applicable)
- [x] I have tested the changes in this PR
- [ ] If I introduced a deprecation:
  - [ ] I followed the [Deprecation Guidelines](./DEPRECATION.md)
  - [ ] I updated the relevant Jira ticket for documentation

---

### 🧪 Testing
<!-- - How it was tested (unit tests, manual, integration) -->  
<!-- - Any special cases covered. -->  

Test is working with IGZ 3.7 env

---

### 🔗 References
- Ticket link: https://iguazio.atlassian.net/browse/ML-11247
- Design docs links:
- External links:

---

### 🚨 Breaking Changes?

- [ ] Yes (explain below)
- [x] No

<!-- If yes, describe what needs to be changed downstream: -->

---

### 🔍️ Additional Notes
<!-- Anything else reviewers should know (follow-up tasks, known issues,
affected areas etc.). -->
<!-- ### 📸 Screenshots / Logs -->