[Pipelines] Experiment name normalization (#8758)
### 📝 Description
<!-- A short summary of what this PR does. -->
<!-- Include any relevant context or background information. -->

When the experiment name of a pipeline doesn't start with the project
name as a prefix `{project}-{experiment}`, then `list_pipelines` doesn't
show the workflow (also in UI). This is handled in the SDK side when
doing `project.run()`.
However a user-provided `experiment name` is accepted in:
- `httpdb.submit_pipeline()`
- `mlrun._run_pipeline` which is widely used in kfp tests

To keep BC, this PR moves that code logic from `project.run()` to
`_create_pipeline` which is used in the endpoint.

Changing this from `project.run()` to the API shouldn't have any effect
because the `name` parameter (the experiment name) is not used anywhere
along the flow after being prefixed with the project name


---

### 🛠️ Changes Made
<!-- - Key changes (e.g., added feature X, refactored Y, fixed Z) -->
- Remove the experiment name normalization from `project.run()`
- Add an experiment name normalization in `_create_pipeline()`

---

### ✅ Checklist
- [ ] I updated the documentation (if applicable)
- [x] I have tested the changes in this PR
- [ ] I confirmed whether my changes are covered by system tests
- [ ] If yes, I ran all relevant system tests and ensured they passed
before submitting this PR
- [ ] I updated existing system tests and/or added new ones if needed to
cover my changes
- [ ] If I introduced a deprecation:
  - [ ] I followed the [Deprecation Guidelines](./DEPRECATION.md)
  - [ ] I updated the relevant Jira ticket for documentation

---

### 🧪 Testing
<!-- - How it was tested (unit tests, manual, integration) -->  
<!-- - Any special cases covered. -->  

Manual + added assertion in existing unit test

---

### 🔗 References
- Ticket link: https://iguazio.atlassian.net/browse/ML-11260
- Design docs links:
- External links:

---

### 🚨 Breaking Changes?

- [ ] Yes (explain below)
- [x] No

<!-- If yes, describe what needs to be changed downstream: -->

---

### 🔍️ Additional Notes
<!-- Anything else reviewers should know (follow-up tasks, known issues,
affected areas etc.). -->
<!-- ### 📸 Screenshots / Logs -->