[Runs] Enable authentication for jobs [feature/ig4-authentication] (#8840)
### 📝 Description
<!-- A short summary of what this PR does. -->
<!-- Include any relevant context or background information. -->
This PR handles cases where a job that runs in MLRun imports the MLRun
SDK, meaning the job itself needs to authenticate with the MLRun API.

Since we no longer use the V3IO access key for authentication, we must
provide an offline token so the SDK can authenticate via Iguazio. To
enable this, we configure environment variables to support
authentication using the offline token.

Specifically, the offline token is passed as the
`MLRUN_AUTH_OFFLINE_TOKEN` environment variable.
Although this is not the recommended approach (the preferred method is
using the igz.yml file), it provides a quick and functional solution
until the proper implementation is completed — see
[ML-11292](https://iguazio.atlassian.net/browse/ML-11292).

In this implementation, we read the offline token from the secret that
belongs to the authenticated user (extracted from the AuthInfo), using
the "default" token name.

Also fixing the bug of `sync_secret_tokens` fails when offline token is
provided via env var and igz.yml is missing
(https://iguazio.atlassian.net/browse/ML-11389)

---

### 🛠️ Changes Made
<!-- - Key changes (e.g., added feature X, refactored Y, fixed Z) -->

---

### ✅ Checklist
- [x] I updated the documentation (if applicable)
- [x] I have tested the changes in this PR
- [ ] I confirmed whether my changes are covered by system tests
- [ ] If yes, I ran all relevant system tests and ensured they passed
before submitting this PR
- [ ] I updated existing system tests and/or added new ones if needed to
cover my changes
- [ ] If I introduced a deprecation:
  - [ ] I followed the [Deprecation Guidelines](./DEPRECATION.md)
  - [ ] I updated the relevant Jira ticket for documentation

---

### 🧪 Testing
<!-- - How it was tested (unit tests, manual, integration) -->  
<!-- - Any special cases covered. -->  

---

### 🔗 References
- Ticket link: fixing also https://iguazio.atlassian.net/browse/ML-11389
- Design docs links:
https://iguazio.atlassian.net/wiki/spaces/ARC/pages/361103361/MLRun+Secret+Tokens+in+IG4
- External links:

---

### 🚨 Breaking Changes?

- [ ] Yes (explain below)
- [x] No
Should be only supported in IG4. 

<!-- If yes, describe what needs to be changed downstream: -->

---

### 🔍️ Additional Notes
<!-- Anything else reviewers should know (follow-up tasks, known issues,
affected areas etc.). -->
<!-- ### 📸 Screenshots / Logs -->


[ML-11292]:
https://iguazio.atlassian.net/browse/ML-11292?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: Katerina Molchanova <35141662+rokatyy@users.noreply.github.com>
Co-authored-by: Amit Elbaz <66309521+elbamit@users.noreply.github.com>