[KubejobRuntime] Secret token mounting for KubejobRuntime (#9013)
### 📝 Description
<!-- A short summary of what this PR does. -->
<!-- Include any relevant context or background information. -->
This PR updates the MLRun API to mount the Kubernetes secret
corresponding to IG4’s offline token `default` (for the running user) as
a file inside the run pod at a predefined folder owned by MLRun. The
folder path is exposed to the pod via an environment variable so that
the SDK running inside the pod can access the token file.

The `ServerSideLauncher._enrich_runtime` method saves the token name in
`run.spec.auth`. The `KubejobRuntimeHandler` then uses this token name
to determine the Kubernetes secret name and mounts it on the runtime
before pod creation.

As a result, `volumes` and `volume_mounts` are added to both the runtime
and pod specifications.

The `MLRUN_AUTH_OFFLINE_TOKEN` env var that was used in
https://github.com/mlrun/mlrun/pull/8840 is removed and not needed
anymore.

---

### 🛠️ Changes Made
<!-- - Key changes (e.g., added feature X, refactored Y, fixed Z) -->
- Removed `MLRUN_AUTH_OFFLINE_TOKEN` and set
`MLRUN_AUTH_WITH_OAUTH_TOKEN__TOKEN_FILE` during auth env enrichment for
pods
- Added an `auth` field to `RunSpec` and to `MLClientCtx`(needed for it
to be saved in DB)
- Find secret's name and mount it as a file to the runtime during
`KubejobRuntimeHandler.run()`

---

### ✅ Checklist
- [ ] I updated the documentation (if applicable)
- [x] I have tested the changes in this PR
- [ ] I confirmed whether my changes are covered by system tests
- [ ] If yes, I ran all relevant system tests and ensured they passed
before submitting this PR
- [ ] I updated existing system tests and/or added new ones if needed to
cover my changes
- [ ] If I introduced a deprecation:
  - [ ] I followed the [Deprecation Guidelines](./DEPRECATION.md)
  - [ ] I updated the relevant Jira ticket for documentation

---

### 🧪 Testing
<!-- - How it was tested (unit tests, manual, integration) -->  
<!-- - Any special cases covered. -->  
- Unit test for mount secret token to runtime and for ensuring IG4 auth
envs are set to runtime
- Manual tests that job pod is deployed and authenticates with mlrun
using the file

---

### 🔗 References
- Ticket link: https://iguazio.atlassian.net/browse/ML-11583
- Design docs links:
https://iguazio.atlassian.net/wiki/spaces/MLRUN/pages/416121541
- External links:

---

### 🚨 Breaking Changes?

- [ ] Yes (explain below)
- [x] No

<!-- If yes, describe what needs to be changed downstream: -->

---

### 🔍️ Additional Notes
<!-- Anything else reviewers should know (follow-up tasks, known issues,
affected areas etc.). -->
<!-- ### 📸 Screenshots / Logs -->

In future PR's, token other than `default` will be allowed and resolved